requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
cot_reports==0.1.3
cron-descriptor==1.4.3
debugpy==1.8.1
decorator==5.1.1
Django==4.2.13
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-stripe-subscription==1.2.1
eventlet==0.36.1
executing==2.0.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.3.7
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.0
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywin32==306
pyzmq==26.0.3
redis==5.0.7
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
stripe==10.1.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.19.0
